Changed add(int, E) method in LinkedAbstractList
One if statement was acting as if it was adding to the back of the list
but was actually adding a node in the second to last spot in the list so
I took out that statement (same as adding to middle of list) and added a
conditional that will actually add to the end of a list (index ==
size()). All tests still pass.